requirements.txt
alabaster==0.7.16
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.2.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.32.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
boto3==1.35.77
botocore==1.35.77
cachetools==5.5.0
Camelot==12.6.29
camelot-py==0.9.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
cryptography==44.0.0
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
docutils==0.20.1
Elixir==0.7.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
executing==2.1.0
fastjsonschema==2.21.0
fqdn==1.5.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imagesize==1.4.1
ipykernel==6.29.5
ipython==8.30.0
isodate==0.7.2
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
json5==0.10.0
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
Levenshtein==0.26.1
llm-confidence==0.2.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
msrest==0.7.1
mypy-extensions==1.0.0
narwhals==1.16.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.1.3
oauthlib==3.2.2
openai==1.57.2
opencv-python==4.10.0.84
openpyxl==3.1.5
orq-ai-sdk==3.0.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pbr==6.1.0
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pdfservices-sdk==4.1.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
poppler-utils==0.1.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
protobuf==5.29.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.25.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.11.0
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.10.4
Send2Trash==1.8.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==7.3.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.5.2
stack-data==0.6.3
streamlit==1.40.2
tabula-py==2.10.0
Tempita==0.6.0
tenacity==9.0.0
terminado==0.18.1
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
XlsxWriter==3.2.0
xlrd==2.0.1
